% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vec.translate.md}
\alias{vec.translate.md}
\title{Translate a vector using a dictionary table in a markdown file.}
\usage{
vec.translate.md(vec, file)
}
\arguments{
\item{vec}{character vector to be translated}

\item{file}{file containing a markdown table with the dictionary}
}
\description{
Translate a vector using a dictionary table in a markdown file.
}
\examples{
vector<-c("France", "Luxbg", "Spain", "Gb")
vec.translate.md(vector, file="dict.md")
}
\seealso{
\code{\link{vec.translate.csv}}, \code{\link{vec.translate.mdstring}}
}

