% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vec.translate.mdstring}
\alias{vec.translate.mdstring}
\title{Translate a vector using a dictionary table in a markdown string}
\usage{
vec.translate.mdstring(vec, dict.md)
}
\arguments{
\item{vec}{character vector to be translated}

\item{dict.md}{string containing a markdown table}
}
\description{
Translate a vector using a dictionary table in a markdown string
}
\examples{
\dontrun{
vec.translate.mdstring(vector)
dict.md="
|from   |to     |
|:------|:------|
|France |France |
|Luxbg  |Luxbg  |
|Spain  |Spain  |
|Gb     |Great Britain     |
"
vec.translate.mdstring(vector, dict.md)
}
}

