% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vec.translate}
\alias{vec.translate}
\title{Translates a vector of characters based on a dictionary}
\usage{
vec.translate(vec, dict, exact = TRUE)
}
\arguments{
\item{vec}{character vector to be translated}

\item{dict}{data frame used as a dictionary based on the first 2 columns}

\item{exact}{whether direct match is done}
}
\description{
Note: Parts of the code copied from Christopher Gandrud's DataCombine package.
}
\examples{
vec<-c("France", "Luxbg", "Spain", "Gb")
dict <- data.frame(from = c("Luxbg", "Gb"),
                   to = c("Luxembourg", "Great Britain"))
vec.translate(vec, dict, exact=TRUE)
}

