% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write.md}
\alias{write.md}
\title{similar as write.csv(), but writes a markdown table into a text file}
\usage{
write.md(df, file = "")
}
\arguments{
\item{df}{a data frame}

\item{file}{a file to write to}
}
\description{
similar as write.csv(), but writes a markdown table into a text file
}

