% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vec.translate.csv}
\alias{vec.translate.csv}
\title{Translate a vector using a dictionary table in a csv file.}
\usage{
vec.translate.csv(vec, file)
}
\arguments{
\item{vec}{character vector to be translated}

\item{file}{file containing a csv with the dictionary}
}
\description{
Translate a vector using a dictionary table in a csv file.
}
\examples{
\dontrun{
vec.translate.csv(vector, file="dict.csv")
}
}

