% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vec.translate.md}
\alias{vec.translate.md}
\title{Translate a vector using a dictionary table in a markdown file.}
\usage{
vec.translate.md(vec, file)
}
\description{
Translate a vector using a dictionary table in a markdown file.
}
\examples{
\dontrun{
vec.translate.md(vector, file="dict.md")
}
}

