% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vec.translate}
\alias{vec.translate}
\title{Function}
\usage{
vec.translate(vec, dict, exact = TRUE)
}
\description{
Note: Parts of the code copied from Christopher Gandrud's DataCombine package.
}
\examples{
\dontrun{
vec<-c("France", "Luxbg", "Spain", "Gb")
dict <- data.frame(from = c("Luxbg", "Gb"),
                   to = c("Luxembourg", "Great Britain"))
vec.translate(vec, dict, exact=T)
}
}

